agritech-news-agent/app/rag/requirements.txt
pymupdf==1.24.10
mistralai==1.2.3
qdrant-client==1.9.1
python-dotenv==1.0.1
numpy==1.26.4
tiktoken==0.7.0
langchain==0.2.12
langchain-text-splitters==0.2.2
